Pushed main GUI code
Working prototype of my code. Immediate areas of improvement is to have an intro screen for choosing the file, better documentation, separation of the scrambling from the GUI, and a better GUI look. I'm currently not satisfied with how it looks since it looks very basic but I think it will have to do for now.

Analysis will be coming by next week or so. For now I want to set up milestones or a Trello for this project so I can organize my goals. I also want to improve my knowledge of Tkinter.